Working Fix scrollable list test Fix get default plugins test
