+++ src/core/layout/scrollable-list/scrollable-list.js
@@ -11,6 +11,7 @@ import { collections } from "../../collections.js";
 import { onScaleChange } from "../../scaler.js";
 import fp from "../../../../lib/lodash/fp/fp.js";
 import { accessibilify } from "../../accessibility/accessibilify.js";
+import {createBackground, resizeBackground } from "./backgrounds.js";
 
 const createPanel = (scene, title, prepTx, parent) => {
     const panel = scene.rexUI.add.scrollablePanel(getConfig(scene, title, prepTx, parent));
@@ -22,49 +23,6 @@ const createPanel = (scene, title, prepTx, parent) => {
 };
 
 export const getType = value => Object.prototype.toString.call(value).slice(8, -1).toLowerCase();
-const nullToUndefined = val => val === null? undefined : val;
-
-
-export const createBackground = {
-    string: (scene, config) => {
-        const background = scene.add.image(0, 0, `${scene.assetPrefix}.${config}`);
-        const safeArea = scene.layout.getSafeArea({}, false);
-
-        debugger
-        background.setScale(safeArea.width / background.width, safeArea.height / background.height);
-        return background;
-    },
-    null: () => {},
-    object: (scene, config) => {
-        const { width, height, x, y } = scene.layout.getSafeArea({}, false);
-        return scene.add.rexNinePatch({
-            x: width / 2 + x,
-            y: height / 2 + y,
-            width,
-            height,
-            key: `${scene.assetPrefix}.${config.key}`,
-            columns: config.columns.map(nullToUndefined),
-            rows: config.rows.map(nullToUndefined),
-        });
-    },
-};
-
-export const resizeBackground = {
-    Image: (background, scene) => {
-        const safeArea = getPanelY(scene);
-        background.setScale(safeArea.width / background.width, safeArea.height / background.height);
-    },
-    null: () => {},
-    NinePatch: (background, scene) => {
-        const { width, height, x, y } = scene.layout.getSafeArea({}, false);
-
-        console.log([width, height, x, y])
-
-        background.x = width / 2 + x;
-        background.y = height / 2 + y;
-        background.resize(width, height);
-    },
-};
 
 const getConfig = (scene, title, prepTx, parent) => {
     const { listPadding: space, assetKeys: keys, assetPrefix } = scene.config;
@@ -221,9 +179,7 @@ export class ScrollableList extends Phaser.GameObjects.Container {
         setupEvents(scene, this.panel);
 
         this.reset = () => {
-            resizePanel(scene, this.panel);
-            console.log(this.background.constructor.name);
-
+            resizePanel(scene, this.panel)();
             resizeBackground[this.background.constructor.name](this.background, scene);
         };
     }
